+++ packages/imperative/__tests__/OverridesLoader.test.ts
@@ -73,7 +73,7 @@ describe("OverridesLoader", () => {
             const packageJson = {
                 name: "host-package",
                 dependencies: {
-                    keytar: "1.0"
+                    "@zowe/secrets-for-zowe-sdk": "1.0"
                 }
             };
 
@@ -98,7 +98,7 @@ describe("OverridesLoader", () => {
             const packageJson = {
                 name: "host-package",
                 dependencies: {
-                    keytar: "1.0"
+                    "@zowe/secrets-for-zowe-sdk": "1.0"
                 }
             };
 
@@ -124,7 +124,7 @@ describe("OverridesLoader", () => {
             const packageJson = {
                 name: "host-package",
                 dependencies: {
-                    keytar: "1.0"
+                    "@zowe/secrets-for-zowe-sdk": "1.0"
                 }
             };
 
@@ -275,7 +275,7 @@ describe("OverridesLoader", () => {
                 // Fake out package.json for the overrides loader
                 const packageJson = {
                     dependencies: {
-                        keytar: "1.0"
+                        "@zowe/secrets-for-zowe-sdk": "1.0"
                     }
                 };
 
@@ -299,7 +299,7 @@ describe("OverridesLoader", () => {
                 // Fake out package.json for the overrides loader
                 const packageJson = {
                     optionalDependencies: {
-                        keytar: "1.0"
+                        "@zowe/secrets-for-zowe-sdk": "1.0"
                     }
                 };
 
@@ -323,7 +323,7 @@ describe("OverridesLoader", () => {
                 // Fake out package.json for the overrides loader
                 const packageJson = {
                     dependencies: {
-                        keytar: "1.0"
+                        "@zowe/secrets-for-zowe-sdk": "1.0"
                     }
                 };