+++ packages/security/src/DefaultCredentialManager.ts
@@ -13,7 +13,7 @@ import { AbstractCredentialManager, SecureCredential } from "./abstract/Abstract
 import { ImperativeError } from "../../error";
 import { Logger } from "../../logger";
 
-import * as keytar from "keytar"; // Used for typing purposes only
+import { keyring as keytar } from "@zowe/secrets-for-zowe-sdk"; // Used for typing purposes only
 
 /**
  * Default Credential Manager is our implementation of the Imperative Credential Manager. This manager invokes methods
@@ -121,11 +121,12 @@ export class DefaultCredentialManager extends AbstractCredentialManager {
             // within our caller's path.
             const requireOpts: any = {};
             if (process.mainModule?.filename != null) {
-                requireOpts.paths = [process.mainModule.filename, ...require.resolve.paths("keytar")];
+                requireOpts.paths = [process.mainModule.filename, ...require.resolve.paths("@zowe/secrets-for-zowe-sdk")];
             }
-            const keytarPath = require.resolve("keytar", requireOpts);
+            // use helper function for require.resolve so it can be mocked in jest tests
+            const keytarPath = require.resolve("@zowe/secrets-for-zowe-sdk", requireOpts);
             Logger.getImperativeLogger().debug("Loading Keytar module from", keytarPath);
-            this.keytar = await import(keytarPath);
+            this.keytar = (await import(keytarPath)).keyring;
         } catch (error) {
             this.loadError = new ImperativeError({
                 msg: `Failed to load Keytar module: ${error.message}`,